fix(supply-chain): adapt shared trade logic to Node and browser loaders
