+++ scripts/shared/comtrade.d.mts
@@ -1,0 +1,1 @@
+export * from './comtrade';
