+++ server/worldmonitor/supply-chain/v1/get-route-impact.ts
@@ -1,4 +1,4 @@
-import { normalizeComtradeProducts } from '../../../../scripts/shared/comtrade-partners.cjs';
+import { normalizeComtradeProducts } from '../../../../scripts/shared/comtrade';
 /**
  * GET /api/supply-chain/v1/get-route-impact
  *
@@ -170,7 +170,8 @@ async function computeImpact(req: GetRouteImpactRequest): Promise<GetRouteImpact
   if (!payload.products?.length) return emptyResponse(req, 'empty');
 
   const normalizedHs2 = String(Number.parseInt(hs2, 10));
-  const matchingHs4s = normalizeComtradeProducts(payload.products).filter((p) => hs4ToHs2(p.hs4) === normalizedHs2);
+  const products = normalizeComtradeProducts(payload.products);
+  const matchingHs4s = products.filter((p) => hs4ToHs2(p.hs4) === normalizedHs2);
   const hs2InSeededUniverse = matchingHs4s.length > 0;
 
   let laneValueUsd = 0;
@@ -190,7 +191,7 @@ async function computeImpact(req: GetRouteImpactRequest): Promise<GetRouteImpact
     }
   }
 
-  const sortedProducts = [...payload.products].sort((a, b) => b.totalValue - a.totalValue);
+  const sortedProducts = [...products].sort((a, b) => b.totalValue - a.totalValue);
   const top5 = sortedProducts.slice(0, 5);
   const topStrategicProducts: StrategicProduct[] = top5.map((p) => ({
     hs4: p.hs4,
